Moving shared functions to common.py
