+++ gstreamer/detect.py
@@ -75,41 +75,10 @@ def generate_svg(src_size, inference_size, inference_box, objs, labels, text_lin
                         fill='none', stroke='red', stroke_width='2'))
     return dwg.tostring()
 
-def make_interpreter(model_file):
-    model_file, *device = model_file.split('@')
-    return tflite.Interpreter(
-      model_path=model_file,
-      experimental_delegates=[
-          tflite.load_delegate(EDGETPU_SHARED_LIB,
-                               {'device': device[0]} if device else {})
-      ])
-
-def input_size(interpreter):
-    """Returns input size as (width, height, channels) tuple."""
-    _, height, width, channels = interpreter.get_input_details()[0]['shape']
-    return width, height, channels
-
-def input_tensor(interpreter):
-    """Returns input tensor view as numpy array of shape (height, width, channels)."""
-    tensor_index = interpreter.get_input_details()[0]['index']
-    return interpreter.tensor(tensor_index)()[0]
-
 def output_tensor(interpreter, i):
     """Returns output tensor view."""
     tensor = interpreter.tensor(interpreter.get_output_details()[i]['index'])()
     return np.squeeze(tensor)
-
-def set_input(interpreter, buf):
-    """Copies data to input tensor."""
-    result, mapinfo = buf.map(Gst.MapFlags.READ)
-    if result:
-        np_buffer = np.reshape(np.frombuffer(mapinfo.data, dtype=np.uint8), input_size(interpreter))
-        input_tensor(interpreter)[:, :] = np_buffer
-        buf.unmap(mapinfo)
-
-def set_interpreter(interpreter, data):
-    set_input(interpreter, data)
-    interpreter.invoke()
 
 class BBox(collections.namedtuple('BBox', ['xmin', 'ymin', 'xmax', 'ymax'])):
     """Bounding box.
